Issue #1 create FileInfo model using core-object
