+++ lib/engines/classic.js
@@ -1,157 +1,79 @@
 var path = require('path');
 var inflection = require('inflection');
+var FileInfo = require('../models/file-info');
 
 var PLURAL_OVERRIDES = {
   'mirage': 'mirage',
   'html': 'html'
 };
+var ClassicFileInfo = FileInfo.extend({
+  plural_overrides: PLURAL_OVERRIDES
+});
 
-function ClassicFileInfo(_options) {
-  var options = _options || {};
-
-  this.sourceRoot = options.sourceRoot;
-  this.destRoot = options.destRoot;
-
-  this.sourceRelativePath = options.sourceRelativePath;
-  this.type = options.type;
-
-  this.ext = options.ext;
-  this.name = options.name;
-  this.collection = options.collection;
-  this.destRelativePath = options.destRelativePath;
-
-  this.populate();
-}
-
-ClassicFileInfo.prototype = {
-  populate: function() {
-    this.populateExt();
-    this.populateName();
-    this.populateCollection();
-    this.populateDestination();
+var ComponentTemplateFileInfo = ClassicFileInfo.extend({
+  init: function(options) {
+    options.collection = 'components';
+    this._super(options);
   },
-
-  populateExt: function() {
-    if (this.ext) { return this.ext; }
-
-    return this.ext = path.extname(this.sourceRelativePath);
-  },
-
   populateName: function() {
-    if (this.name) {
-      return this.name;
-    }
+    this._super();
 
-    var pathParts = this.sourceRelativePath.split('/');
-    var rootFolder = pathParts[0];
-
-    var name = this.sourceRelativePath
-          .replace(new RegExp('^' + rootFolder + '/'), '') // remove leading type dir
-          .replace(new RegExp(this.ext + '$'), '') // remove extension
-          .replace(new RegExp('/' + this.type + '$'), ''); // remove trailing type
+    var name = this.name.replace(/^components\//, '');
 
     return this.name = name;
-  },
+  }
+});
+var MainFileInfo = ClassicFileInfo.extend({
+  init: function(options) {
+    options.type = 'main';
 
-  populateCollection: function() {
-    if (this.collection) {
-      return this.collection;
+    if (options.sourceRelativePath === 'app.js') {
+      options.name = 'main';
     }
+    this._super(options);
 
-    var collection;
-
-    switch (this.type) {
-    case 'adapter':
-    case 'serializer':
-    case 'model':
-      collection = 'models';
-      break;
-
-    case 'route':
-    case 'controller':
-    case 'template':
-      collection = 'routes';
-      break;
-
-    default:
-      collection = inflection.pluralize(this.type, PLURAL_OVERRIDES[this.type]);
-    }
-
-    return this.collection = collection;
+    this.collection = 'main';
   },
-
   populateDestination: function() {
     if (this.destRelativePath) {
       return this.destRelativePath;
     }
 
-    var destRelativePath = this.collection + '/' + this.name + '/' + this.type + this.ext;
+    var destRelativePath = this.name + this.ext;
 
     return this.destRelativePath = destRelativePath;
   }
-};
+});
 
-function ComponentTemplateFileInfo(options) {
-  options.collection = 'components';
+var SingleTypeCollectionFileInfo = ClassicFileInfo.extend({
+  populateDestination: function() {
+    if (this.destRelativePath) {
+      return this.destRelativePath;
+    }
 
-  ClassicFileInfo.apply(this, arguments);
-}
-ComponentTemplateFileInfo.prototype = Object.create(ClassicFileInfo.prototype);
-ComponentTemplateFileInfo.prototype.constructor = ComponentTemplateFileInfo;
-ComponentTemplateFileInfo.prototype.$superPopulateName = ClassicFileInfo.prototype.populateName;
+    var destRelativePath = this.collection + '/' + this.name + this.ext;
 
-ComponentTemplateFileInfo.prototype.populateName = function() {
-  if (this.name) {
-    return this.name;
+    return this.destRelativePath = destRelativePath;
   }
+});
 
-  var name = this.$superPopulateName()
-        .replace(/^components\//, '');
+var MixinFileInfo = ClassicFileInfo.extend({
+  init: function(options) {
+    options.type = 'util';
 
-  return this.name = name;
-};
+    this._super(options);
+  },
+  populateDestination: function() {
+    if (this.destRelativePath) {
+      return this.destRelativePath;
+    }
 
-function MainFileInfo(options) {
-  options.type = 'main';
+    var destRelativePath = this.collection + '/mixins/' + this.name + this.ext;
 
-  if (options.sourceRelativePath === 'app.js') {
-    options.name = 'main';
+    return this.destRelativePath = destRelativePath;
   }
-
-  ClassicFileInfo.apply(this, arguments);
-
-  this.collection = 'main';
-}
-MainFileInfo.prototype = Object.create(ClassicFileInfo.prototype);
-MainFileInfo.prototype.constructor = MainFileInfo;
-
-MainFileInfo.prototype.populateDestination = function() {
-  if (this.destRelativePath) {
-    return this.destRelativePath;
-  }
-
-  var destRelativePath = this.name + this.ext;
-
-  return this.destRelativePath = destRelativePath;
-};
-
-function SingleTypeCollectionFileInfo() {
-  ClassicFileInfo.apply(this, arguments);
-}
-SingleTypeCollectionFileInfo.prototype = Object.create(ClassicFileInfo.prototype);
-SingleTypeCollectionFileInfo.prototype.constructor = SingleTypeCollectionFileInfo;
-
-SingleTypeCollectionFileInfo.prototype.populateDestination = function() {
-  if (this.destRelativePath) {
-    return this.destRelativePath;
-  }
-
-  var destRelativePath = this.collection + '/' + this.name + this.ext;
-
-  return this.destRelativePath = destRelativePath;
-};
-
-SingleTypeCollectionFileInfo.typeIncluded = function(type) {
+});
+function typeIncluded(type) {
   switch (type) {
     // data
   case 'adapter':
@@ -172,25 +94,7 @@ SingleTypeCollectionFileInfo.typeIncluded = function(type) {
   default:
     return true;
   }
-};
-
-function MixinFileInfo(options) {
-  options.type = 'util';
-
-  ClassicFileInfo.apply(this, arguments);
 }
-MixinFileInfo.prototype = Object.create(ClassicFileInfo.prototype);
-MixinFileInfo.prototype.constructor = MixinFileInfo;
-
-MixinFileInfo.prototype.populateDestination = function() {
-  if (this.destRelativePath) {
-    return this.destRelativePath;
-  }
-
-  var destRelativePath = this.collection + '/mixins/' + this.name + this.ext;
-
-  return this.destRelativePath = destRelativePath;
-};
 
 module.exports = {
   buildFor: function(filePath, options) {
@@ -219,7 +123,7 @@ module.exports = {
 
     if (topLevelDirectory === 'mixins') {
       return new MixinFileInfo(options);
-    } else if (SingleTypeCollectionFileInfo.typeIncluded(options.type)) {
+    } else if (typeIncluded(options.type)) {
       return new SingleTypeCollectionFileInfo(options);
     } else {
       return new ClassicFileInfo(options);
